requirements.txt
numpy==1.19.2
Pillow==7.2.0
PyOpenGL==3.1.5
